Move lmdb tests to own file
This change moves the lmdb integration tests to their own file.  This
should improve readability of the submodules.
